Add friends. Friends list and chat list don't work

+++ js/src/main/scala/ionictabs/client/domain/Chat.scala
@@ -4,10 +4,10 @@ import scala.scalajs.js
 import scala.scalajs.js.annotation.JSExportAll
 
 @JSExportAll
-class Chat(val id: Int, val name: String, val lastText: String, val face: String)
+class Chat(val id: Long, val name: String, val lastText: String, val face: String)
 
 object Chat {
   val unknown: Chat = new Chat(-1, "", "", "")
 
-  def apply(id: Int, name: String, lastText: String, face: String) = new Chat(id, name, lastText, face)
+  def apply(id: Long, name: String, lastText: String, face: String) = new Chat(id, name, lastText, face)
 }
